Improved error handling and output
Should now identify more problem files (e.g. cassette ends in wrong
orientation).  This allows the code to run to the end of the sequence
file.

Now sorts results based on the frequency and reports the number of files
that threw an error.

Next job is to go through the incorrect sequence results one at a time
to try and identify and resolve why they were wrong (mostly they are out
by a couple of bp).